Separate views file into smaller files.
Fix validate_ticket_view tests.

Move get_chapter_events to its own file.

Remove uneccesary imports

Move start_payment to separate file.

Move reserve_ticket to separate file.

Move get_session_payment_status to separate file.

Remove views.py
