+++ backend/bittan/bittan/views/__init__.py
@@ -1,0 +1,6 @@
+from bittan.views.validate_ticket import validate_ticket
+from bittan.views.get_chapter_events import get_chapter_events
+from bittan.views.start_payment import start_payment
+from bittan.views.reserve_ticket import reserve_ticket
+from bittan.views.get_session_payment_status import get_session_payment_status
+
